+++ home/components/OpenFromClipboardButton.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Keyboard, Linking, Platform } from 'react-native';
+import { Clipboard, Keyboard, Linking, Platform } from 'react-native';
 
 import UrlUtils from '../utils/UrlUtils';
 import ListItem from './ListItem';
@@ -10,6 +10,13 @@ type Props = {
 };
 
 export default class OpenFromClipboardButton extends React.Component<Props> {
+  onPress = async () => {
+    const contents = await Clipboard.getString();
+    if (UrlUtils.conformsToExpoProtocol(contents)) {
+      this.openAppUrl(contents);
+    }
+  };
+
   render() {
     const { clipboardContents, isValid } = this.props;
 
@@ -17,9 +24,10 @@ export default class OpenFromClipboardButton extends React.Component<Props> {
     if (!isValid) {
       return (
         <ListItem
+          onPress={this.onPress}
           subtitle={
             Platform.OS === 'ios'
-              ? 'Press ⌘+v to move clipboard to simulator.'
+              ? 'Press ⌘+v to move clipboard to simulator. Tap to open.'
               : 'Project URLs on your clipboard will appear here.'
           }
           last
@@ -39,10 +47,14 @@ export default class OpenFromClipboardButton extends React.Component<Props> {
   }
 
   private handlePressAsync = async () => {
+    this.openAppUrl(this.props.clipboardContents);
+  };
+
+  private openAppUrl(clipboardUrl: string) {
     // note(brentvatne): navigation should do this automatically
     Keyboard.dismiss();
 
-    const url = UrlUtils.normalizeUrl(this.props.clipboardContents);
+    const url = UrlUtils.normalizeUrl(clipboardUrl);
     Linking.canOpenURL(url) && Linking.openURL(url);
-  };
+  }
 }
